[tired:ui] Enh | Only rewrite intermediate file, if explicitly stated

+++ tired/ui.py
@@ -14,12 +14,14 @@ def select_yn(title="") -> bool:
     return bool(selected_option_id)  # bool hack: 0 and 1 match False and True
 
 
-def get_input_using_temporary_file(file_path=".tmp", editor="vim", initial_message=""):
+def get_input_using_temporary_file(file_path=".tmp", editor="vim", initial_message="", force_rewrite_initial_message=True):
     import tired.command
+    import os
 
     # Create file
-    with open(file_path, 'w') as f:
-        f.write(initial_message)
+    if not os.path.isfile(file_path) or force_rewrite_initial_message:
+        with open(file_path, 'w') as f:
+            f.write(initial_message)
 
     tired.command.execute(f"{editor} {file_path}")
